% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_plot}
\alias{annotate_plot}
\title{plot subplot}
\usage{
annotate_plot(.sub_plot, .x, .y, .size)
}
\arguments{
\item{.sub_plot}{plot}

\item{.x}{place}

\item{.y}{place}

\item{.size}{size}
}
\description{
make subplot
}
