% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ncbi.R
\name{extract_ncbi}
\alias{extract_ncbi}
\title{Extract data from NCBI xml file.}
\usage{
extract_ncbi(.xml, .id)
}
\arguments{
\item{.xml}{xml file}

\item{.id}{id file}
}
\description{
This function make index_1.fasta and index_2.fasta from SampleSheet.csv.
}
