% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_num.R
\name{funguild_DB}
\alias{funguild_DB}
\title{Creating text file for FUNGuild.}
\usage{
funguild_DB(table, combined, combined_col_names = T)
}
\arguments{
\item{table}{sumary table like "summary_nonchimeras.txt"}

\item{combined}{identifying result like "combined.txt"}

\item{combined_col_names}{logical. Default is TRUE. This parameter determins ignoreing col_names of combined.txt or not.}
}
\description{
If your identifying data (e.g, "combined.txt") dose not have col_names, this function automatically name col_names (taxonomic levels).
So, you should carefully check creating col_names is appropriate or not.
}
\examples{
example
}
