% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_index.R
\name{index_make}
\alias{index_make}
\title{Making Index fast file for claident.}
\usage{
index_make(.filename, .skip = 18)
}
\arguments{
\item{.filename}{"SampleSheet.csv"}

\item{.skip}{number of skip row, default 18}
}
\description{
This function make index_1.fast and index_2.fast from SampleSheet.csv.
}
