% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_time.R
\name{interval_time}
\alias{interval_time}
\title{Cariculate interval time}
\usage{
interval_time(df)
}
\arguments{
\item{df}{A data.frame}
}
\description{
Cariculate intercal time from camera survey.
}
\examples{
interval_time
}
