% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabr.R
\name{label_extract}
\alias{label_extract}
\title{Title of the simple function to extract_num.}
\usage{
label_extract(string)
}
\arguments{
\item{string}{A character}
}
\description{
The second paragraph is recognized as the description.
It is not recommended to use @title and @description explicitly.
}
\examples{
extract_num(ABD_2-1-10_GG)
}
