% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_unique_number.R
\name{make_unique_number}
\alias{make_unique_number}
\title{make unique number.}
\usage{
make_unique_number(.interval, .pattern = "T+F|T+|.")
}
\arguments{
\item{.interval}{interval time.}

\item{.pattern}{reglex.}
}
\description{
make unique number.
}
