% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_tbl.R
\name{matrix_to_tbl}
\alias{matrix_to_tbl}
\title{Convert matrix to tibble.}
\usage{
matrix_to_tbl(.mat, .row, .rowname = NULL)
}
\arguments{
\item{.mat}{matrix}

\item{.row}{logical, convert row to column.}

\item{.rowname}{character, rowname.}
}
\description{
Convert matrix to tibble.
}
