% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subplot.R
\name{my_sub_plot}
\alias{my_sub_plot}
\title{Make sub plot}
\usage{
my_sub_plot(.data, .aes, .stat, .pie)
}
\arguments{
\item{.data}{data}

\item{.aes}{aes}

\item{.stat}{stat}

\item{.pie}{logical, make pie chart.}
}
\description{
Make sub plot
}
