% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmds_ggplot2.R
\name{nmds_ggplot2}
\alias{nmds_ggplot2}
\title{Parce output of metaMDS for ggplot2.}
\usage{
nmds_ggplot2(.data, .key, .sample, .nest = F, .col_name = "mds")
}
\arguments{
\item{.data}{data}

\item{.key}{rowname}

\item{.sample}{sample data.}

\item{.nest}{logical}

\item{.col_name}{column name to extract}
}
\description{
Parce output of metaMDS for ggplot2.
}
