% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_clustered_otu.R
\name{parse_clustered_otu}
\alias{parse_clustered_otu}
\title{parse_clustered_otu}
\usage{
parse_clustered_otu(.clustered_otu, .col_name = X1)
}
\arguments{
\item{.clustered_otu}{clustered.otu.gz produced by Claident.}

\item{.col_name}{name of column.}
}
\description{
parse_clustered_otu
}
\examples{
clustered_otu <- data.frame(
   otu = c(">representative_otu1", "assigned_otu1", "assigned_otu2",
          ">representative_otu2", "assigned_otu1",
          ">representative_otu3")
)
parse_clustered_otu(.clustered_otu = clustered_otu,
                    .col_name = otu)
}
