% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_GBQualifier.R
\name{parse_GBQualifier}
\alias{parse_GBQualifier}
\title{Parse GBQualifier}
\usage{
parse_GBQualifier(.GBQualifier = list())
}
\arguments{
\item{.GBQualifier}{GBQualifier}
}
\description{
This function use internally parse_xml.
}
\examples{
GBFeature_quals <- list(GBQualifier = list(
                   "GBQualifier_name" = "organism",
                   "GBQualifier_value" = "Ceratobasidiaceae sp."))

parse_GBQualifier(.GBQualifier = GBFeature_quals)
}
