% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_xml.R
\name{parse_xml}
\alias{parse_xml}
\title{Prepare xml}
\usage{
parse_xml(.xml_list)
}
\arguments{
\item{.xml_list}{xml as_list()}

\item{.manual}{logical}
}
\description{
Prepare xml
}
\examples{
\dontrun{
# load file.
xml <- (read_xml("test.xml")

# prepare
by_sample <- prepare(.xml = xml)

# make table.
result <- map_dfr(.x = by_sample, .f = parse_xml, .id = "id") \%>\%
  pivot_wider(names_from = "name", values_from = "value")
}
}
