% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_table.R
\name{PCR_table_96}
\alias{PCR_table_96}
\title{Make table for PCR 96 plate.}
\usage{
PCR_table_96(.data, .col_label, .plate_num)
}
\arguments{
\item{.data}{data.frame.}

\item{.col_label}{column names}

\item{.plate_num}{96 plate number}
}
\description{
Make table for PCR 96 plate.
}
