% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posiible_parse_GBQualifier.R
\name{posiible_parse_GBQualifier}
\alias{posiible_parse_GBQualifier}
\title{Parse GBQualifier (possibly).}
\usage{
posiible_parse_GBQualifier(...)
}
\arguments{
\item{.GBQualifier}{GBQualifier}
}
\description{
This function use internally in parse_xml.
}
\examples{
GBFeature_quals <- list(GBQualifier = list(
                   "GBQualifier_name" = "organism",
                   "GBQualifier_value" = "Ceratobasidiaceae sp."))

parse_GBQualifier(.GBQualifier = GBFeature_quals)
}
