% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_NA_pluck.R
\name{return_NA_pluck}
\alias{return_NA_pluck}
\title{return_NA_pluck}
\usage{
return_NA_pluck(.x, ..., .default = NA)
}
\arguments{
\item{.x}{list}

\item{...}{element of list}

\item{.default}{return NA}
}
\description{
return_NA_pluck
}
