% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_variables.R
\name{set_formulas}
\alias{set_formulas}
\title{Creating formula.}
\usage{
set_formulas(
  .first_expl_var,
  .other_expl_var,
  .resp_var,
  .type = "lm",
  .interaction
)
}
\arguments{
\item{.first_expl_var}{first explanatry variable}

\item{.other_expl_var}{explanatry variables}

\item{.resp_var}{responese variable}

\item{.type}{model type}

\item{.interaction}{interaction}
}
\description{
Creating formula for fitting multiple models.
}
