% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_cmpl.R
\name{str_cmpl}
\alias{str_cmpl}
\title{Complement character.}
\usage{
str_cmpl(
  .id,
  .gather = 4,
  .fill = "0",
  .prefix = NULL,
  .suffix = NULL,
  .sep = ""
)
}
\arguments{
\item{.id}{character}

\item{.gather}{max num}

\item{.fill}{filling character.}

\item{.prefix}{prefix}

\item{.suffix}{suffix}

\item{.sep}{separator.}
}
\description{
Complement character.
}
\examples{
str_cmpl(.id = 2)
str_cmpl(.id = "Apple", .gather = 10, .fill = "A")
str_cmpl(.id = 34, .gather = 4, .fill = "9")
str_cmpl(.id = 12, .gather = 4, .fill = "0",
         .prefix = "suetsugulab", .sep = "-")
str_cmpl(.id = 12, .gather = 4, .fill = "0",
         .prefix = "suetsugulab-", .suffix = "A")
}
