% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_to_fasta.R
\name{table_to_fasta}
\alias{table_to_fasta}
\title{Convert table to fasta.}
\usage{
table_to_fasta(.data, .otu, .seq)
}
\arguments{
\item{.data}{data.frame}

\item{.otu}{otu column name.}

\item{.seq}{sequence column name.}
}
\description{
Convert table to fasta.
}
