% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_to_matrix.R
\name{tbl_to_matrix}
\alias{tbl_to_matrix}
\title{Convert tibble to matrix.}
\usage{
tbl_to_matrix(.tbl, .row = F, .rowname = NULL)
}
\arguments{
\item{.tbl}{tibble}

\item{.row}{logical. Want rownames ?}

\item{.rowname}{character.}
}
\description{
Convert tibble to matrix.
}
