% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_lag.R
\name{time_lag}
\alias{time_lag}
\title{alculate time lag}
\usage{
time_lag(.time, .col, .fill, .order = "top")
}
\arguments{
\item{.time}{time}

\item{.col}{col}

\item{.fill}{logical}

\item{.order}{character, "top" or "end".}
}
\description{
alculate time lag
}
