% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funguild.R
\name{Funguild_DB}
\alias{Funguild_DB}
\title{Creating table for FUNGuild.}
\usage{
Funguild_DB(.otu_table, .identify_list)
}
\arguments{
\item{.otu_table}{sumary table.}

\item{.identify_list}{identifying result.}
}
\description{
Creating table for analysis of FUNGuild ( http://www.stbates.org/guilds/app.php ).
This function requir two tables, (1) summary_table of OTU/read_num, (2) identification of each OTU.
Names of output file is Funguild_DB.txt.
}
