% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funguild_DB.R
\name{Funguild_DB}
\alias{Funguild_DB}
\title{Creating table for FUNGuild.}
\usage{
Funguild_DB(.otu_table, .identify_list)
}
\arguments{
\item{.otu_table}{sumary table.}

\item{.identify_list}{identifying result.}
}
\description{
Creating table for analysis of FUNGuild.
}
