% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ncbi_xml.R
\name{load_ncbi_xml}
\alias{load_ncbi_xml}
\title{laod fungal sequence and sample data from NCBI using accession ID}
\usage{
load_ncbi_xml(.id, .filename)
}
\arguments{
\item{.id}{character, Genbank accession ID.}

\item{.filename}{saved xml file name.}
}
\description{
laod fungal sequence and sample data from NCBI using accession ID
}
