% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_summary_table.R
\name{make_summary_table}
\alias{make_summary_table}
\title{Make summary table}
\usage{
make_summary_table(
  .id,
  .funguild,
  .seq,
  .sample_info,
  .tax_pat = "unite",
  .taxon,
  .remove = T
)
}
\arguments{
\item{.id}{siz002}

\item{.funguild}{funguild}

\item{.seq}{clustered.fasta}

\item{.sample_info}{sample information}

\item{.tax_pat}{UNITE DB or overall_genus or another.}

\item{.taxon}{manual}

\item{.remove}{remove size.}
}
\description{
This function make summary tabel.
}
